+++ FbxParser/FbxNodeList.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
+using FbxTools.Internal;
 
 namespace FbxTools
 {
